refactor(DataType): 重构所有 id 数据类型

+++ src/main/java/vip/gadfly/chakkispring/common/interceptor/AuthorizeVerifyResolverImpl.java
@@ -68,7 +68,7 @@ public class AuthorizeVerifyResolverImpl implements AuthorizeVerifyResolver {
         UserDO user = LocalUser.getLocalUser();
         if (verifyAdmin(user))
             return true;
-        long userId = user.getId();
+        Integer userId = user.getId();
         String permission = meta.getPermission();
         String module = meta.getModule();
         List<PermissionDO> permissions = userService.getUserPermissions(userId);
