+++ src/main/java/vip/gadfly/chakkispring/mapper/ClassMapper.java
@@ -19,7 +19,7 @@ public interface ClassMapper extends BaseMapper<ClassDO> {
      * @param userId 用户id
      * @return 所有班级
      */
-    List<ClassDO> selectUserClasses(@Param("userId") Long userId);
+    List<ClassDO> selectUserClasses(@Param("userId") Integer userId);
 
     /**
      * 获得用户的所有班级id
@@ -27,7 +27,7 @@ public interface ClassMapper extends BaseMapper<ClassDO> {
      * @param userId 用户id
      * @return 所有班级id
      */
-    List<Long> selectUserClassIds(@Param("userId") Long userId);
+    List<Integer> selectUserClassIds(@Param("userId") Integer userId);
 
     /**
      * 通过id获得班级个数
@@ -35,7 +35,7 @@ public interface ClassMapper extends BaseMapper<ClassDO> {
      * @param id id
      * @return 个数
      */
-    int selectCountById(@Param("id") Long id);
+    int selectCountById(@Param("id") Integer id);
 
     // /**
     //  * 检查用户是否在该名称的班级里
@@ -43,7 +43,7 @@ public interface ClassMapper extends BaseMapper<ClassDO> {
     //  * @param userId    用户id
     //  * @param className 班级名
     //  */
-    // int selectCountUserByUserIdAndClassName(@Param("userId") Long userId, @Param("className") String className);
+    // int selectCountUserByUserIdAndClassName(@Param("userId") Integer userId, @Param("className") String className);
 
     /**
      * 根据学期id和教师id查询班级列表
@@ -52,7 +52,7 @@ public interface ClassMapper extends BaseMapper<ClassDO> {
      * @param teacherId  教师id
      * @return 班级列表
      */
-    List<ClassDO> selectClassesBySemesterAndTeacher(Long semesterId, Long teacherId);
+    List<ClassDO> selectClassesBySemesterAndTeacher(Integer semesterId, Integer teacherId);
 
     /**
      * 根据学期id和学生id查询班级列表
@@ -61,5 +61,5 @@ public interface ClassMapper extends BaseMapper<ClassDO> {
      * @param userId     学生id
      * @return 班级列表
      */
-    List<ClassDO> selectClassesBySemesterAndStudent(Long semesterId, Long userId);
+    List<ClassDO> selectClassesBySemesterAndStudent(Integer semesterId, Integer userId);
 }
