+++ src/main/java/vip/gadfly/chakkispring/mapper/GroupMapper.java
@@ -20,7 +20,7 @@ public interface GroupMapper extends BaseMapper<GroupDO> {
      * @param userId 用户id
      * @return 所有分组
      */
-    List<GroupDO> selectGroupsByUserId(@Param("userId") Long userId);
+    List<GroupDO> selectGroupsByUserId(@Param("userId") Integer userId);
 
     /**
      * 获得用户的所有分组id
@@ -28,7 +28,7 @@ public interface GroupMapper extends BaseMapper<GroupDO> {
      * @param userId 用户id
      * @return 所有分组id
      */
-    List<Long> selectUserGroupIds(@Param("userId") Long userId);
+    List<Integer> selectUserGroupIds(@Param("userId") Integer userId);
 
     /**
      * 通过id获得分组个数
@@ -36,7 +36,7 @@ public interface GroupMapper extends BaseMapper<GroupDO> {
      * @param id id
      * @return 个数
      */
-    int selectCountById(@Param("id") Long id);
+    int selectCountById(@Param("id") Integer id);
 
     /**
      * 检查用户是否在该名称的分组里
@@ -44,5 +44,5 @@ public interface GroupMapper extends BaseMapper<GroupDO> {
      * @param userId    用户id
      * @param groupName 分组名
      */
-    int selectCountUserByUserIdAndGroupName(@Param("userId") Long userId, @Param("groupName") String groupName);
+    int selectCountUserByUserIdAndGroupName(@Param("userId") Integer userId, @Param("groupName") String groupName);
 }
