+++ src/main/java/vip/gadfly/chakkispring/mapper/StudentClassMapper.java
@@ -15,5 +15,5 @@ public interface StudentClassMapper extends BaseMapper<StudentClassDO> {
 
     int insertBatch(@Param("relations") List<StudentClassDO> relations);
 
-    int removeByClassId(@Param("class_id") Long class_id);
+    int removeByClassId(@Param("class_id") Integer class_id);
 }
