+++ src/main/java/vip/gadfly/chakkispring/module/message/WsHandler.java
@@ -46,7 +46,7 @@ public interface WsHandler {
      * @param message 要发送的消息
      * @throws IOException 发送io异常
      */
-    void sendMessage(Long userId, TextMessage message) throws IOException;
+    void sendMessage(Integer userId, TextMessage message) throws IOException;
 
     /**
      * 发送消息
@@ -55,7 +55,7 @@ public interface WsHandler {
      * @param message 要发送的消息
      * @throws IOException 发送io异常
      */
-    void sendMessage(Long userId, String message) throws IOException;
+    void sendMessage(Integer userId, String message) throws IOException;
 
     /**
      * 发送消息
@@ -81,7 +81,7 @@ public interface WsHandler {
      * @param message 消息
      * @throws IOException 异常
      */
-    void broadCastToGroup(Long groupId, String message) throws IOException;
+    void broadCastToGroup(Integer groupId, String message) throws IOException;
 
     /**
      * 广播
@@ -98,7 +98,7 @@ public interface WsHandler {
      * @param message 文本消息
      * @throws IOException 异常
      */
-    void broadCastToGroup(Long groupId, TextMessage message) throws IOException;
+    void broadCastToGroup(Integer groupId, TextMessage message) throws IOException;
 
     /**
      * 处理会话异常
